remove applyFix util
